services/auth/requirements.txt
# Core Dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
email-validator==2.2.0
pyotp==2.9.0  # For 2FA

# Validation
pydantic==2.9.2
pydantic[email]==2.9.2
pydantic-settings==2.5.2

# Redis
redis==5.0.8
aioredis==2.0.1

# Vault Integration
hvac==2.3.0

# HTTP Client
httpx==0.27.2
aiohttp==3.10.5

# CSRF Protection
itsdangerous==2.2.0

# Rate Limiting
slowapi==0.1.9

# Monitoring
prometheus-client==0.20.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0

# Utilities
python-dateutil==2.9.0
pytz==2024.2
aiofiles==24.1.0

# Logging
structlog==24.4.0
python-json-logger==2.0.7